pyproject.toml
[project]
name = "data-jobs"
version = "0.1.0"
description = "Retrieves job postings from job boards and analyses them"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "wordcloud>=1.9.4",
]
